Remove logging handlers from library code
It is not recommended for libraries to have their own logging handlers, as this should be set on the application side. Setting handlers on library code can cause issues on the application side, such as when `logging.config.dictConfig `is called more than once. Moving logging.yml to test directory and setting handlers for unit testing only. Users have to set their own logging handlers in case they want to use logging functionality.