Refactor rm_none option implementation
Stop using regex and use a parser function caching instead. This faster than the previous implementation and allows the first field to be None (which wasn't allowed before). However, now the decorated user-defined class must be a named tuple (or implement a _asdict method that returns a dictionary version of its contents) and therefore a minor API break.